refactor: run public web search through Effect
