+++ src/web-search.test.ts
@@ -1,9 +1,12 @@
+import { Effect } from "effect";
 import { expect, test } from "vitest";
 import { createPublicWebSearchTool, performWebSearch } from "./web-search";
 
+const runSearch = (...args: Parameters<typeof performWebSearch>) => Effect.runPromise(performWebSearch(...args));
+
 test("normalizes Cloudflare Web Search results and sends only the query upstream", async () => {
   const requests: string[] = [];
-  const result = await performWebSearch("Cloudflare Workers", {
+  const result = await runSearch("Cloudflare Workers", {
     fetch: async (query) => {
       requests.push(query);
       return {
@@ -29,7 +32,7 @@ test("normalizes Cloudflare Web Search results and sends only the query upstream
 });
 
 test("caps public web results and marks the response truncated", async () => {
-  const result = await performWebSearch("Cloudflare", {
+  const result = await runSearch("Cloudflare", {
     maxResults: 3,
     fetch: async () => ({
       items: Array.from({ length: 6 }, (_, index) => ({
@@ -50,7 +53,7 @@ test("caps public web results and marks the response truncated", async () => {
 });
 
 test("drops result URLs that are not absolute http or https citations", async () => {
-  const result = await performWebSearch("citations", {
+  const result = await runSearch("citations", {
     fetch: async () => ({
       items: [
         { title: "Relative", url: "/article", description: "Not a citation" },
@@ -69,12 +72,12 @@ test("drops result URLs that are not absolute http or https citations", async ()
 
 test("returns safe errors without exposing upstream credentials", async () => {
   const secret = "test-web-search-secret";
-  const failed = await performWebSearch("failure", {
+  const failed = await runSearch("failure", {
     fetch: async () => {
       throw new Error(`Authorization: Bearer ${secret}`);
     },
   });
-  const unavailable = await performWebSearch("missing", {});
+  const unavailable = await runSearch("missing", {});
   const tool = createPublicWebSearchTool();
   const toolResult = await tool.execute({ query: "missing binding" }, { env: {} } as any);
